requirements.txt
streamlit>=1.32.0
supabase>=2.4.0
google-generativeai>=0.8.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
